Un-factors errcheck and adds for nearly everything
Turns out that the only potential user of an errcheck other than
less-than-zero is genders_handle_create, which cannot use the
errnum/errormsg functions anyway. So this simplifies it back to a single
errcheck function and adds to to nearly everything else. (The
error-processing functions also seem unable to use it meaningfully.)